Created ID3 Algorithm and Decision Tree
The ID3 algorithm is working well with the example dataset from the pdf
on canvas. It uses an abstract data entry class. The basic structure of
the decision tree is made, but much more still needs to be added. A
simple testing function will test the algorithm and sample dataset
("testData.txt").